refactor(idea_service): extract scoring helpers (#163)
  before: 2160
  after:  2080 (-80)
